+++ linkerd/protocol/http/src/main/scala/io/buoyant/linkerd/protocol/HttpConfig.scala
@@ -105,6 +105,10 @@ case class HttpConfig(
     ResponseClassifiers.NonRetryableServerFailures orElse super.baseResponseClassifier
 
   @JsonIgnore
+  override def responseClassifier =
+    ResponseClassifiers.NonRetryableChunked(super.responseClassifier)
+
+  @JsonIgnore
   override val protocol: ProtocolInitializer = HttpInitializer
 
   @JsonIgnore